requirements.txt
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.5
defusedxml==0.7.1
Django==5.0.2
django-crispy-forms==2.1
django-ranged-response==0.2.0
django-recaptcha==4.0.0
django-resized==1.0.2
django-simple-captcha==0.6.0
django-tinymce==4.0.0
idna==3.6
oauthlib==3.2.2
pillow==10.3.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python-slugify==8.0.4
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.10.0
urllib3==2.2.1
